+++ CHANGELOG.md
@@ -24,6 +24,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Fixed
 
+- Desired RX1 delay and desired beacon frequency not being possible to set for OTAA devices.
+
 ### Security
 
 - Fix open redirect vulnerability for Console/Account App logins.