+++ CHANGELOG.md
@@ -48,6 +48,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Removed
 
 - The device version identifiers no longer have the `serial_number`, `vendor_id` and `vendor_profile_id` fields.
+- Automatic migrations of the Network Server database using `ns-db migrate` from versions prior to v3.23 are removed. Migrating from prior versions should be done through v3.23 instead.
 
 ## [3.23.2] - 2023-01-18